android fix reported bugs
