+++ android/app/release/output-metadata.json
@@ -21,12 +21,16 @@
     {
       "minApi": 28,
       "maxApi": 30,
-      "baselineProfiles": ["baselineProfiles/1/app-release.dm"]
+      "baselineProfiles": [
+        "baselineProfiles/1/app-release.dm"
+      ]
     },
     {
       "minApi": 31,
       "maxApi": 2147483647,
-      "baselineProfiles": ["baselineProfiles/0/app-release.dm"]
+      "baselineProfiles": [
+        "baselineProfiles/0/app-release.dm"
+      ]
     }
   ],
   "minSdkVersionForDexing": 24
