--- android/app/src/main/res/drawable/logo.xml
@@ -1,16 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="1000dp"
-    android:height="1000dp"
-    android:viewportWidth="1000"
-    android:viewportHeight="1000">
-  <group>
-    <clip-path
-        android:pathData="M0,0h1000v1000h-1000z"/>
-    <path
-        android:pathData="M1000,0H0V1000H1000V0Z"
-        android:fillColor="#FF7701"/>
-    <path
-        android:pathData="M467.4,158.2C506.4,132.2 558.6,160.1 558.6,206.9V264.1C558.6,296.4 584.8,322.7 617.2,322.7H730.5C762.8,322.7 789.1,348.9 789.1,381.2V623.4C789.1,652 774.8,678.8 751.1,694.7L535.5,839.9C496.6,866.1 444.2,838.2 444.2,791.3V727.8C444.2,695.4 418,669.2 385.6,669.2H271.5C239.1,669.2 212.9,643 212.9,610.6V374.4C212.9,345.7 227.2,318.9 251,303L467.4,158.2ZM384.7,454.6C364.2,454.6 351.6,476.9 362.1,494.5C424.1,597.8 574.3,597.6 638.2,495.5C649.3,477.8 636.7,454.6 615.8,454.6H384.7Z"
-        android:fillColor="#ffffff"/>
-  </group>
-</vector>
