+++ android/app/src/main/res/values/colors.xml
@@ -3,6 +3,6 @@
     <color name="black">#FF000000</color>
     <color name="white">#FFFFFFFF</color>
 
-    <color name="brand">#FFFF7701</color>
+    <color name="brand">#FF5C00</color>
     <color name="brand_dark">#FFE56A00</color>
 </resources>
